# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ACG Theogony API",
    "description": "ACG × 神话原型 关系知识图谱 API",
    "version": "2.0.0"
  },
  "paths": {
    "/api/characters": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "List Characters",
        "operationId": "list_characters_api_characters_get",
        "parameters": [
          {
            "name": "mythology",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mythology"
            }
          },
          {
            "name": "className",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classname"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gender"
            }
          },
          {
            "name": "hasRelation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasrelation"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterListItem"
                  },
                  "title": "Response List Characters Api Characters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/characters/{character_id}": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Get Character",
        "operationId": "get_character_api_characters__character_id__get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/{character_id}/relationships": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Character Relationships",
        "operationId": "character_relationships_api_characters__character_id__relationships_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          },
          {
            "name": "includePending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includepending"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelationshipDTO"
                  },
                  "title": "Response Character Relationships Api Characters  Character Id  Relationships Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/{character_id}/family": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Character Family",
        "operationId": "character_family_api_characters__character_id__family_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/{character_id}/ego": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Character Ego",
        "operationId": "character_ego_api_characters__character_id__ego_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          },
          {
            "name": "hops",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "default": 1,
              "title": "Hops"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/games/daily/guess": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Daily Guess",
        "operationId": "daily_guess_api_games_daily_guess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyGuess"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CharacterDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "wikiId": {
            "type": "integer",
            "title": "Wikiid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases",
            "default": []
          },
          "className": {
            "type": "string",
            "title": "Classname",
            "default": ""
          },
          "prototype": {
            "type": "string",
            "title": "Prototype",
            "default": ""
          },
          "mythology": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mythology"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "mythologyBackground": {
            "type": "string",
            "title": "Mythologybackground",
            "default": ""
          },
          "alignment": {
            "type": "string",
            "title": "Alignment",
            "default": ""
          },
          "gender": {
            "type": "string",
            "title": "Gender",
            "default": ""
          },
          "imageUrl": {
            "type": "string",
            "title": "Imageurl",
            "default": ""
          },
          "detailUrl": {
            "type": "string",
            "title": "Detailurl",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "fgo"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra",
            "default": {}
          },
          "degree": {
            "type": "integer",
            "title": "Degree",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "wikiId",
          "name"
        ],
        "title": "CharacterDTO"
      },
      "CharacterListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "className": {
            "type": "string",
            "title": "Classname",
            "default": ""
          },
          "mythology": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mythology"
          },
          "imageUrl": {
            "type": "string",
            "title": "Imageurl",
            "default": ""
          },
          "degree": {
            "type": "integer",
            "title": "Degree",
            "default": 0
          },
          "hasRelation": {
            "type": "boolean",
            "title": "Hasrelation",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CharacterListItem"
      },
      "Confidence": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "verified"
        ],
        "title": "Confidence"
      },
      "DailyGuess": {
        "properties": {
          "guess": {
            "type": "string",
            "minLength": 1,
            "title": "Guess"
          }
        },
        "type": "object",
        "required": [
          "guess"
        ],
        "title": "DailyGuess"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RelationOrigin": {
        "type": "string",
        "enum": [
          "manual",
          "llm",
          "user",
          "wiki"
        ],
        "title": "RelationOrigin"
      },
      "RelationshipDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "sourceId": {
            "type": "string",
            "title": "Sourceid"
          },
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "type": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "directed": {
            "type": "boolean",
            "title": "Directed",
            "default": true
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence",
            "default": ""
          },
          "origin": {
            "$ref": "#/components/schemas/RelationOrigin"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus",
            "default": "approved"
          },
          "sourceName": {
            "type": "string",
            "title": "Sourcename",
            "default": ""
          },
          "targetName": {
            "type": "string",
            "title": "Targetname",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "sourceId",
          "targetId",
          "type",
          "confidence",
          "origin"
        ],
        "title": "RelationshipDTO"
      },
...
      "RelationshipType": {
        "type": "string",
        "enum": [
          "PARENT_OF",
          "SIBLING_OF",
          "SPOUSE_OF",
          "LOVER_OF",
          "MASTER_OF",
          "ALLY_OF",
          "ENEMY_OF",
          "FOUGHT_WITH",
          "MENTOR_OF",
          "PROTOTYPE_IS",
          "GENDER_SWAP",
          "DERIVED_FROM",
          "COMPOSITE_OF",
          "BELONGS_TO"
        ],
        "title": "RelationshipType",
        "description": "关系类型（存储层只用规范方向，逆关系在入库时归一化）。"
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "approved",
          "pending",
          "rejected"
        ],
        "title": "ReviewStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}